Restructured ui styles
Moved the styles to the companion `tsx` file, instead of having them in the styles folder, for easier maintenance
