+++ electron/src/ui/App.tsx
@@ -1,5 +1,4 @@
 import { AppFooter, AppHeader, AppMain } from './layout';
-import './styles/App.css';
 
 function App() {
   return (
