+++ electron/src/ui/layout/AppMain.tsx
@@ -1,4 +1,4 @@
-import styles from '../styles/AppMain.module.css';
+import styles from './AppMain.module.css';
 
 export function AppMain() {
   return (
