--- electron/src/ui/styles/App.css
@@ -1,1 +1,0 @@
-@reference "./index.css";
